# Optimize `reduced_vocabulary`